Fix POST exercise route and sanitizeDate()
In this commit, I fixed the route /api/users/:_id/exercises.

There was an issue where FCC expected the date to be in the format
(EEE MM dd YYYY). This was not possible to do with the default
Date alone because every time the date was parsed, the parsed date
would be off by 1 day. This was because Date() parses strings and
sets them up in UTC time. The strings provided for this project in
question were expected to follow the ISO8601 format. I resorted to
using Luxon to help me parse out the date in the correct format.

Luxon was used on sanitizeDate() to parse out the date and output
a date in the correct format. If no input is provided to it, it
returns the current date in the ISO8601 format.

Another issue I was running into was that the test for the route
expected an object to be returned with an ID. Normally, one would
expect the return value to include the value of the subdocument
that was just created. Instead, the test was expecting the ID of the
username associated with the exercise. The output of the test was
not clear at all with what it expected, and I had to do a lot of
digging to find out the cause. This has been resolved altogether,
and having implemented sanitizeDate() will help me to implement
the route to retrieve a user's logs.